1: Implements jitter in train loop and allows noise added to training data
FunctionDataset now supports generating data with noise baked in, instead of
adding noise randomly on reads.

Also, generally tweaks model (hyper)parameters in an effort to make it harder for
the standard model to converge. The idea is that if the standard model can't easily
converge, jitter and preconditioning will be able to stand out more.