Create most of code to run study on LP initialized models
Currently, the new notebook can create an LP initialized model, train
it, and return training/test results per epoch very easily. The study will
utilize the `train_many` function to do this for a given number of models,
and collect all the results.

The CollectedResults class is unfinished, but will be written to easily store
and analyze the reuslts of the study.